requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
pandas>=1.5.0
scikit-image>=0.19.0
tqdm>=4.64.0
